% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alE}
\alias{alE}
\title{Arc length estimation.}
\usage{
alE(x, q1, q2, dc, type)
}
\arguments{
\item{x}{A vector of sample values.}

\item{q1, q2}{Vectors specifying the quantiles over which arc length segments are to be computed.}

\item{dc}{TRUE/FALSE:  Should the discrete or continuous sample statistic be used.}

\item{type}{The type of bandwidth estimator for the underlying KDE; see \code{\link{bw}}.}
}
\value{
A list with the following components (see \code{\link{optim}}):
\itemize{
\item par: The estimated parameters.
\item abstol: The absolute tolarence level (default 1e-15).
\item fail: An integer code indicating convergence.
\item fncount: Number of function evaluations.
}
}
\description{
Estimate distributional parameters using the method of arc lengths.
}
\details{
This method is currently only implimented for the normal distribution.  The underlying C code for the Nelder-Mead method of the optim function is used for optimising the objective function.  The tolarence level is set at 1e-15, and a maximum number of 1000 iterations is allowed.
}
\examples{
x <- rnorm(1000)
alE(x,0.025, 0.975, TRUE, -1)
alE(x,c(0.025, 0.5), c(0.5, 0.975), TRUE, -1)
alE(x,0.025, 0.975, FALSE, -1)
alE(x,c(0.025, 0.5), c(0.5, 0.975), FALSE, -1)

}

