% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bw}
\alias{bw}
\title{Kernel density bandwidth parameters.}
\usage{
bw(x, type)
}
\arguments{
\item{x}{A vector of data points.}

\item{type}{An integer:
\itemize{
\item 1: Silverman's rule of thumb.
\item 2: Silverman's adapted rule of thumb.
\item default: Silvermans rule of thumb (number 1).
}}
}
\value{
Bandwidth estimator based on Silverman's rule of thumb.
}
\description{
Calculate the bandwidth estimator using various methods.
}
\examples{
set.seed(1)
x <- rnorm(100)
bw(x, 1)
bw(x, 2)

}

