% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixBoot.R
\name{fixBoot}
\alias{fixBoot}
\title{Trimmed bootstrap distributions.}
\usage{
fixBoot(object, trim = 0, na.rm = TRUE)
}
\arguments{
\item{object}{An \code{alKDEboot}, \code{mmKDEboot}, \code{kappa4alBoot}, or \code{kappa4nlsBoot}.}

\item{trim}{#' @param trim The fraction (0 to 0.5) of observations to be trimmed from each end of the jackknife distribution before the mean and standard deviation are computed. Values of \code{trim} outside that range are taken as the nearest endpoint.}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
The \code{object} is returned with the estimated coefficients corresponding to the trimmed bootstrap distribution.
}
\description{
Remove bootstrap estimates associated with extreme bootstrap distribution values, resulting from arc length methods.
}
