% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixJack.R
\name{fixJack}
\alias{fixJack}
\title{Trimmed jackknife distributions.}
\usage{
fixJack(object, trim, na.rm = TRUE)
}
\arguments{
\item{object}{An \code{alKDEjack}, \code{mmKDEjack}, \code{kappa4alJack}, or \code{kappa4nlsJack}.}

\item{trim}{#' @param trim The fraction (0 to 0.5) of observations to be trimmed from each end of the jackknife distribution before the mean and standard deviation are computed. Values of \code{trim} outside that range are taken as the nearest endpoint.}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
The \code{object} is returned with the estimated coefficients corresponding to the trimmed bootstrap distribution.
}
\description{
Remove jackknife estimates associated with extreme jackknife distribution values, resulting from arc length methods.
}

