% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dGPD}
\alias{dGPD}
\alias{pGPD}
\alias{qGPD}
\alias{rGPD}
\title{The Generalised Pareto Distribution (GPD).}
\usage{
dGPD(x, mu, sigma, alpha)

pGPD(q, mu, sigma, alpha)

qGPD(p, mu, sigma, alpha)

rGPD(n, mu, sigma, alpha)
}
\arguments{
\item{x, q}{A data point, or quantile, at which the GPD should be evaluated (may be vectors).}

\item{mu}{The location parameter of the GPD.}

\item{sigma}{The scale parameter of the GPD (should be strictly positive).}

\item{alpha}{The shape parameter of the GPD (valid for all real numbers).}

\item{p}{A probability, at which the GPD should be evaluated.}

\item{n}{The number of random numbers which should be generated for the GPD.}
}
\value{
dGPD: The estimated value of the density function of the GPD at the point \code{x}.

pGPD: The value of the cumulative distribution function of the GPD at the point \code{x}.

qGPD: The \code{x}th quantile of the GPD.

rGPD: \code{n} random numbers from the GPD.
}
\description{
Density, distribution function, quantile function and random generation for the generalised Pareto distribution (GPD) with location parameter \code{mu}, scale parameter \code{sigma}, and shape parameter \code{alpha}. .
}
\examples{
library(alR)
dGPD(0.5, 0, 1, 2)
pGPD(0.5, 0, 1, 2)
qGPD(0.5, 0, 1, 2)
rGPD(10, 0, 1, 2)
}
