% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GPDMLE}
\alias{GPDMLE}
\alias{GPDqML}
\alias{GPDMSE}
\title{The qML Method for the Generalised Pareto Distribution (GPD).}
\usage{
GPDMLE(x)

GPDqML(x)

GPDMSE(x)
}
\arguments{
\item{x}{A vector of data points.}
}
\value{
GPDMLE:  A named vector containing the maximum likelihood estimates for the GPD fitted to the data \code{x}.

GPDqML:  A named vector containing the estimated parameters for the GPD fitted to the data \code{x}, using quasi-maximum likelihood.

GPDMSE:  A named vector containing the estimated parameters for the GPD fitted to the data \code{x}, using maximum spacings estimation (MSE).
}
\description{
Estimation of the parameters of the generalised Pareto distribution (GPD) with location parameter \code{mu}, scale parameter \code{sigma}, and shape parameter \code{alpha}.  using the quasi-maximum-likelihood method.
}
\examples{
x <- rGPD(100, 0, 1, 2)
GPDMLE(x)
GPDqML(x)
\dontrun{
GPDMSE(x)
}
}
