% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kappa4NLSobj}
\alias{kappa4NLScon}
\alias{kappa4NLSheq}
\alias{kappa4NLShin}
\alias{kappa4NLSobj}
\title{Sigmoidal curve fitting.}
\usage{
kappa4NLSobj(parms, xvec, y, x_min, x_max)

kappa4NLScon(parms, xvec, y, x_min, x_max)

kappa4NLShin(parms, xvec, y, x_min, x_max)

kappa4NLSheq(parms, xvec, y, x_min, x_max)
}
\arguments{
\item{parms}{A numeric vector of parameters to be estimated.}

\item{xvec}{A numeric vector of independent observations.}

\item{y}{A numeric vector of dependent observations.}

\item{x_min}{The minimum xvec value.}

\item{x_max}{The maximum xvec value.}
}
\value{
kappa4NLSobj: The nonlinear least squares objective function.

kappa4NLScon: A vector with three conditions evaluated.

kappa4NLShin: A vector specifying a single nonlinear inequality constraint.

kappa4NLSheq: A vector specifying two nonlinear equality constraints.
}
\description{
Support functions for fitting four-parameter kappa sigmoidal curves.
}

