% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lulu}
\alias{lulu}
\title{LULU smoother.}
\usage{
lulu(x)
}
\arguments{
\item{x}{A real-valued vector.}
}
\value{
The LULU-smoothed version of \code{x}.
}
\description{
Performs LULU smoothing of the provided vector.
}
\examples{
x <- rnorm(10)
lulu(x)
}

