% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmKDEshort.R
\name{mmKDEshort}
\alias{mmKDEshort}
\title{Moment matching for kernel density estimators.}
\usage{
mmKDEshort(formula, data = list(), lower, upper, itermax, ...)
}
\arguments{
\item{formula}{An LHS ~ RHS formula, specifying the linear model to be estimated.}

\item{data}{A data.frame which contains the variables in \code{formula}.}

\item{lower, upper}{Numeric vectors, of length equal to the number of independent variables, for the lower and upper bounds for the parameters to be estimated.}

\item{itermax}{Number of iterations for the Differential Evolution algorithm.}

\item{...}{Arguments to be passed on to \code{DEoptim.control()} of the Differential Evolution algorithm.}
}
\value{
mmKDEshort: A list with the following components:
\itemize{
\item coefficients: A vector of estimated coefficients.
\item error: The value of the objective function.
}
}
\description{
Estimate parameters of a linear model by matching the moments of kernel density estimators.
}
\details{
A shortened version of \code{mmKDE}.
}
\examples{
x <- 1:10
y <- x+rnorm(10)
mmKDEshort(y~x, lower=c(-2,2), upper=c(2,2), itermax=50)

}

