% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{momKDE}
\alias{momKDE}
\title{Objective function for KDE moment matching.}
\usage{
momKDE(beta, gamma, momy, kdeGaussMom)
}
\arguments{
\item{beta}{Vector of regression parameters.}

\item{gamma}{Design matrix.}

\item{momy}{A numerical vector of moments to be matched to (same length as \code{beta}).}

\item{kdeGaussMom}{A R function object to be passed that calculates exact moments.}
}
\value{
Sum of squared differences between \code{gamma}*\code{beta} and \code{momy}.
}
\description{
The moments of kernel density estimates are matched.
}

