% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pairSort}
\alias{pairSort}
\title{Sorted vector index.}
\usage{
pairSort(x)
}
\arguments{
\item{x}{A real-valued vector to be sorted.}
}
\value{
A list with two components:
\itemize{
\item sorted: The sorted version of \code{x}.
\item index: The index of the \eqn{i^{th}} element in \code{x}.
}
}
\description{
The sorted vector is returned along with the original index of the vector it belonged to.
}
\examples{
pairSort(c(5, 2, 6))
}
