% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qsamp}
\alias{qsamp}
\title{Empirical sample quantile.
Calculate empirical sample quantile.}
\usage{
qsamp(x, q)
}
\arguments{
\item{x}{A numeric vector, specifying the sample for which the quantile is to be calculated.}

\item{q}{A real number between 0 and 1 inclusive, specifying the desired quantile.}
}
\value{
The empirical quantile of the provided sample.
}
\description{
Empirical sample quantile.
Calculate empirical sample quantile.
}
\examples{
x<-rnorm(100)
qsamp(x, 0.5)
}
