% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Silverman}
\alias{Silverman}
\title{Silverman's rule of thumb.}
\usage{
Silverman(x)
}
\arguments{
\item{x}{A vector of data points.}
}
\value{
Bandwidth estimator based on Silverman's rule of thumb.
}
\description{
Calculate the bandwidth estimator using Silverman's rule of thumb.
}
\examples{
set.seed(1)
x <- rnorm(100)
Silverman(x)

}

