% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Silverman2}
\alias{Silverman2}
\title{Silverman's adapted rule of thumb.}
\usage{
Silverman2(x)
}
\arguments{
\item{x}{A vector of data points.}
}
\value{
Bandwidth estimator based on Silverman's adapted rule of thumb.
}
\description{
Calculate the bandwidth estimator using Silverman's adapted rule of thumb.
}
\examples{
set.seed(1)
x <- rnorm(100)
Silverman2(x)

}

