% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alR-package.R
\docType{package}
\name{alR}
\alias{alR}
\alias{alR-package}
\title{alR arc lengths of statistical functions.}
\description{
Functions using arc lengths, for performing estimation, regression, and classification are provided.
}
\author{
m. Theodor Loots <theo.loots@gmail.com>
}

